fallback to Query, Subscription and Mutation type
Summary:
If the schema doesn't define a query, subscription, or mutation type, we default somewhere in the JS compiler to Query, Mutation, and Subscription.
This implements this fallback in the schema and removes the "server-extensions.graphql" file that would conflict if the schema SDL actually does define the schema types differently.

Reviewed By: alunyov

Differential Revision: D19912147

fbshipit-source-id: 3ccb0ce73fa13df7a7dd4721ba16f7e0d1bb2763