Refactor store release buffer management
Summary:
This is a small but subtle refactor to how the release buffer is managed:
* Maintain the invariant that the release buffer only contains entries with a refCount of 0 (technically, it contains the ids of those entries).
* This implies adding items to the release buffer when their refCount becomes 0, and removing them when their refCount goes from 0->1.
* This requires guarding against multiple dispose() calls for a given retain() call - ensuring a 1:1 correspondence makes the above 0<->1 checks correct.

This has a few benefits. Notably, a query that was disposed and re-retained would previously have stayed in the release buffer, causing the buffer to appear more full than it really was. This could cause GC to run earlier than necessary.

Reviewed By: jstejada

Differential Revision: D20002411

fbshipit-source-id: 13456312224cdf6eca30e50a6c488f6050d0b603