Change toPromise to complete Observable
Summary:
## Background

The `Observable::toPromise()` method has a surprising behavior: it returns a Promise that resolves when the source Observable yields its first value (or if the Observable completes without yielding any values), but **it also cancels the source Observable once it has resolved the Promise**.

In practice, `toPromise()` is most frequently used in combination as `fetchQuery().toPromise()` as a way to start a request and wait for it to complete. Such code was generally written before the advent of Incremental Delivery and Data-Driven Dependencies (3D), both of which change the contract of a Relay request from being 1 response to N. Canceling a request after yielding only the first value means that any remaining work won't be processed.

Note that RxJS has a different behavior, which is to resolve the promise only when the source Observable completes, and to yield the *last* value of that Observable: https://github.com/Reactive-Extensions/RxJS/blob/master/doc/api/core/operators/topromise.md (outdated, i couldn't find solid documentation for the latest version).

## Proposed Solution

The proposed solution here is to continue resolving the promise with the *first* value yielded by the promise, but to not cancel the source Observable (allow it to run to completion). This means that the values that consuming code will receive doesn't change, and instead they may see additional side effects. However, again, most usages of this API are such that either:

* The request did not use Incremental Delivery or 3D and would have finished after the first value anyway...
* ...or the request did use those features are was probably accidentally throwing away the extra data (in which case, this change is a fix).

## Alternate Solutions

* Migrate callsites to a different API with a different contract, possibly deprecating the existing API.
  * Upside: avoids changing behavior for anyone using RelayObservable *not* for the simpler cases of calling `fetchQuery`.
  * Downside: more work
* Change the behavior to be closer to that of RxJS: resolve the Promise with the *last* yielded value, only when the source Observable completes.
  * Downside: could return different data, which could cause crashes for code that assumed that would always get a root GraphQL payload and now could end up with an incremental payload instead.

Reviewed By: jstejada

Differential Revision: D20816597

fbshipit-source-id: 07d5612754097df78d46b23888a4cc15f41cfb38