handle field transform
Summary:
This and the next diff implement handle fields for reader nodes.

In JS this is implemented as the single `FieldHandleTransform`. `FieldHandleTransform` does 2 things in JS:
1. Rename the alias and server field of fields with handles to point to a virtual field in the store.
2. Remove some/all arguments based on `filters`/`dynamicKey` of the handle.

In our Rust compiler, we cannot do step 1 in a transform since the IR points at the field definition in the schema where these types don't exist, so I split this into 2 parts:

- The previous diff (D20860338) updates the Relay codegen to rename the "schema name" of fields (linked and scalar) to the handle field in the store.
- **This diff** is related. It implements the part that we can from the `FieldHandleTransform` as a transform: removing filtered args and adding the `__dynamicKey`.

## Alternative

An other option could be to do all of it in the codegen to keep the code more together, but blow up the codegen more and with more complicated logic of filtering the args.

Reviewed By: jstejada

Differential Revision: D20860508

fbshipit-source-id: cc5a414fcee1bbde9486678f013562d816265de6